update db script

+++ exam-server/file/db/10_t_question_judge.sql
@@ -7,7 +7,7 @@ create table t_question_judge(
                                is_del bit(1) not null default true comment '删除标识：0-未删除；1-已删除',
                                create_time timestamp DEFAULT CURRENT_TIMESTAMP NOT NULL COMMENT '创建时间',
                                update_time timestamp DEFAULT CURRENT_TIMESTAMP on update CURRENT_TIMESTAMP NOT NULL COMMENT '更新时间',
-                               creator int(20) COMMENT '创建者',
-                               updater int(20) COMMENT '更新者',
+                               creator bigint(20) COMMENT '创建者',
+                               updater bigint(20) COMMENT '更新者',
                                version int(5) DEFAULT 0 NOT NULL COMMENT '版本号，乐观锁专用'
 ) default charset = utf8 comment '判断题';
